+++ dwds/lib/dart_web_debug_service.dart
@@ -85,7 +85,6 @@ class Dwds {
     bool emitDebugEvents = true,
     bool isInternalBuild = false,
     Future<bool> Function()? isFlutterApp,
-    String basePath = '',
   }) async {
     globalLoadStrategy = loadStrategy;
     isFlutterApp ??= () => Future.value(true);
@@ -151,7 +150,6 @@ class Dwds {
       injected,
       spawnDds,
       launchDevToolsInNewWindow,
-      basePath,
     );
 
     return Dwds._(